Prevent flash of bg blur on page load
Since we were waiting for doc ready to do our scroll check, but using
"reverse" styles (eg: :not(.at-scroll-top)) there was a moment in time
where the bg was visible, and blurred, before we swooped in and removed
the blur.

Although the classname "scrolled-past-top" isn't quite as logically
clear, it's more useful for our purpose. And we avoid jumping thourgh
hoops with a "doc-ready" class.